port par command, and move some common functionality over to compat.h for even easier porting
